code cleanup
Former-commit-id: a4c34c0438bb26603b4cefc31a7419936df37887

+++ src/ves/xplorer/scenegraph/rtt/Unit.cxx
@@ -166,7 +166,7 @@ const Unit::InputToUniformMap& Unit::GetInputToUniformMap() const
     return mInputToUniformMap;
 }
 ////////////////////////////////////////////////////////////////////////////////
-osg::Texture* Unit::GetInputTexture( int inputIndex ) const
+osg::Texture* Unit::GetInputTexture( size_t inputIndex ) const
 {
     if( inputIndex < mInputTextures.size() )
     {
